+++ packages/number-line/configure/src/defaults.js
@@ -1,31 +1,38 @@
-export default {
-  model: {
-    prompt: 'Question Prompt goes here.',
-    correctResponse: [],
-    graph: {
-      domain: [-5, 5],
-      initialElements: [],
-      showMinorTicks: true,
-      initialType: 'PF',
-      exhibitOnly: false,
-      availableTypes: {
-        PF: true,
-        PE: false,
-        LFF: false,
-        LEF: false,
-        LFE: false,
-        LEE: false,
-        RFN: false,
-        RFP: false,
-        REN: false,
-        REP: false
-      }
-    }
-  },
-  configuration: {
-    prompt: {
-      settings: true,
-      label: 'Prompt'
-    }
+export const model = {
+  correctResponse: [],
+  graph: {
+    domain: { min: -5, max: 5 },
+    ticks: {
+      minor: 1,
+      major: 2
+    },
+    arrows: {
+      left: true,
+      right: true
+    },
+    maxNumberOfPoints: 20,
+    width: 500,
+    height: 400,
+    initialType: "PF",
+    exhibitOnly: false,
+    availableTypes: {
+      PF: true,
+      LFF: true,
+      LEF: true,
+      LFE: true,
+      LEE: true,
+      RFN: true,
+      RFP: true,
+      REN: true,
+      REP: true
+    },
+    initialElements: []
+  }
+};
+
+export const configuration = {
+  prompt: {
+    settings: true,
+    label: "Prompt"
   }
 };