+++ packages/point-intercept/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.1.0...@pie-element/point-intercept@4.1.1) (2019-06-23)
+
+**Note:** Version bump only for package @pie-element/point-intercept
+
+
+
+
+
 # [4.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/point-intercept@4.0.8...@pie-element/point-intercept@4.1.0) (2019-06-21)